Separate async job submission
