+++ client/src/components/Tool/submit/index.js
@@ -1,0 +1,1 @@
+export { submitToolJob } from "./submitAsync";
